added map editor, saves to temp.yml for now

+++ src/resource_manager.rb
@@ -11,9 +11,12 @@ class ResourceManager
   
   def load_map(map_name)
     map = YAML::load_file("#{MAP_PATH}#{map_name}.yml")
-    map.script = MapScript.new(File.open("#{MAP_PATH}#{map_name}.rb").readlines.join("\n"))
     map.resource_manager = self
-    map.script.map = map
+
+    if File.exists? "#{MAP_PATH}#{map_name}.rb"
+      map.script = MapScript.new(File.open("#{MAP_PATH}#{map_name}.rb").readlines.join("\n"))
+      map.script.map = map
+    end
     map
   end
